fix(接口): 星级换为小数
星级字段换位小数

+++ housekeeping-admin/housekeeping-admin-interface/src/main/java/com/housekeeping/admin/entity/EmployeesDetails.java
@@ -45,7 +45,7 @@ public class EmployeesDetails extends Model<EmployeesDetails> {
     private String tags;    /* 標簽 */
     private String recordOfFormalSchooling; /* 学历 */
     private Boolean blacklistFlag;    /* 是否加入黑名單 */
-    private Integer starRating;     /* 星级 0 1 2 3 4 5 6 7 8 9 10 */
+    private Float starRating;     /* 星级 0 1 2 3 4 5 */
     private LocalDateTime createTime;    /* 创建时间 */
     private LocalDateTime updateTime;    /* 修改时间 */
     private Integer lastReviserId;    /* 最后修改人 */
